pyproject.toml
[project]
name = "agentic-intelligence-workflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Michelle Espinoza",email = "michelleespinoza1@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "pandasai (>=3.0.0,<4.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "langchain-openai (>=1.1.11,<2.0.0)",
    "langgraph (>=1.1.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "telethon (>=1.42.0,<2.0.0)",
    "yfinance (>=1.2.0,<2.0.0)",
    "typer (>=0.24.1,<0.25.0)",
    "slack-bolt (>=1.27.0,<2.0.0)",
    "geopandas (>=1.1.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
